Can now read in parameters files. There are three ways to set parameters: directly in the algorithm.parameters.params dictionary, in a new parameters text file, or by passing parameters in from the command line. Means problems can be re-run exactly by simply reading in the params file.
Also changed params['PROBLEM'] to params['FITNESS_FUNCTION'], as that's essentially exactly what it does. You specify the desired fitness function for a problem. This means you will need a unique fitness function for different problems, but it gets rid of fitness wheels with large "if, else" statements that previously set the correct fitness functions. It is now much easier to create new problems and to expand ponyge, provided the current framework is adhered to.