Fix rich comparison functions of Individual
In the cases that the fitness function has been set to minimize, an
individual with NaN fitness was considered be greater than another
individual. This can never be the case since if the fitness of an
individual is NaN, that is not a correct evolution and hence, it cannot
be greater than the other individual.